+++ tests/test_commands/conftest.py
@@ -232,3 +232,38 @@ def small_ts():
     df = pd.DataFrame({"segment": ["segment_0"], "timestamp": [pd.Timestamp("2020-01-01")], "target": [1]})
     df = TSDataset.to_dataset(df=df)
     return TSDataset(df=df, freq="D")
+
+
+@pytest.fixture
+def base_timeseries_numeric_segments_path():
+    df = pd.DataFrame(
+        {
+            "timestamp": list(pd.date_range("2021-06-01", periods=100)) * 2,
+            "target": np.arange(200),
+            # segments with numeric names and leading zeros
+            "segment": ["01234"] * 100 + ["12345"] * 100,
+        }
+    )
+    tmp = NamedTemporaryFile("w")
+    df.to_csv(tmp, index=False)
+    tmp.flush()
+    yield Path(tmp.name)
+    tmp.close()
+
+
+@pytest.fixture
+def base_timeseries_numeric_segments_exog_path():
+    df_regressors = pd.DataFrame(
+        {
+            "timestamp": list(pd.date_range("2021-06-01", periods=120)) * 2,
+            "regressor_1": np.arange(240),
+            "regressor_2": np.arange(240) + 5,
+            # segments with numeric names and leading zeros
+            "segment": ["01234"] * 120 + ["12345"] * 120,
+        }
+    )
+    tmp = NamedTemporaryFile("w")
+    df_regressors.to_csv(tmp, index=False)
+    tmp.flush()
+    yield Path(tmp.name)
+    tmp.close()